Update gaussian_elim.py
changed from "doc.py"  import "function" because function 
"row" was moved to rowops.py. Added "rowscale_0" to imports even though it is not used. Maybe needed to reduce "size" rows of matrices when possible, ie, if R1 = [0, -3, 6], then we can do R1 = rowscale_0(R1, 1/3) to get R1 = [0, -1, 2]? This is important for reducing machine error? to be explored...